Created two new functions: updateTodayProject() and updateWeekProject()
updateTodatProject() will go through each project and check if the tasks are due that current day and move them into the today project.

updateWeekProject() will go through each prject to check if the task is due this week and will add them to the This Week project and sort them.